+++ cmd/ozon-tenant-api/main.go
@@ -2,8 +2,8 @@ package main
 
 import (
 	"fmt"
+	"github.com/ozoncp/ocp-tenant-api/internal/tenant"
 	"github.com/ozoncp/ocp-tenant-api/internal/utils-lesson2"
-	"github.com/ozoncp/ocp-tenant-api/internal/utils-lesson3"
 )
 
 func main() {
@@ -93,13 +93,13 @@ func lesson2RemoveElements() {
 func lesson3() {
 	fmt.Println("Результаты третьего урока")
 	lesson3createTenant()
-	utils_lesson3.OpenFileInLoop()
+	tenant.OpenFileInLoop()
 	lesson3ToMapWithId()
 	lesson3SplitToButch()
 }
 
 func lesson3createTenant() {
-	newTenant := utils_lesson3.Tenant{
+	newTenant := tenant.Tenant{
 		Id:   1,
 		Name: "First",
 		Type: 1,
@@ -107,21 +107,21 @@ func lesson3createTenant() {
 	fmt.Println("Создан первый елемент типа Tenant: " + newTenant.ToString())
 }
 
-func lesson3GetTenantSlice() []utils_lesson3.Tenant {
-	data := make([]utils_lesson3.Tenant, 3)
-	newTenant1 := utils_lesson3.Tenant{
+func lesson3GetTenantSlice() []tenant.Tenant {
+	data := make([]tenant.Tenant, 3)
+	newTenant1 := tenant.Tenant{
 		Id:   1,
 		Name: "First",
 		Type: 1,
 	}
 	data[0] = newTenant1
-	newTenant2 := utils_lesson3.Tenant{
+	newTenant2 := tenant.Tenant{
 		Id:   2,
 		Name: "Second",
 		Type: 2,
 	}
 	data[1] = newTenant2
-	newTenant3 := utils_lesson3.Tenant{
+	newTenant3 := tenant.Tenant{
 		Id:   3,
 		Name: "End",
 		Type: 3,
@@ -134,14 +134,14 @@ func lesson3ToMapWithId() {
 	fmt.Println("Тестируем ToMapWithId")
 	data := lesson3GetTenantSlice()
 
-	mapWithTenant, err := utils_lesson3.ToMapWithId(data)
+	mapWithTenant, err := tenant.ToMapWithId(data)
 	check(err)
 	for key, value := range mapWithTenant {
 		fmt.Printf("key: %d; value: %s\n", key, value.ToString())
 	}
 
 	data[2].Id = 1
-	_, err2 := utils_lesson3.ToMapWithId(data)
+	_, err2 := tenant.ToMapWithId(data)
 	if err2.Error() == "Tenant::Id is not uniq" {
 		fmt.Println("Тестирование ToMapWithId успешно")
 	} else {
@@ -157,13 +157,13 @@ func lesson3SplitToButch() {
 		fmt.Printf("index: %d; value: %s\n", index, value.ToString())
 	}
 	fmt.Println("Разделить на кусочки по одному:")
-	for index, value := range utils_lesson3.SplitToButch(data, 1) {
+	for index, value := range tenant.SplitToButch(data, 1) {
 		for index2, value2 := range value {
 			fmt.Printf("index1: %d; index2: %d; value: %s\n", index, index2, value2.ToString())
 		}
 	}
 	fmt.Println("Разделить на кусочки по двое:")
-	for index, value := range utils_lesson3.SplitToButch(data, 2) {
+	for index, value := range tenant.SplitToButch(data, 2) {
 		for index2, value2 := range value {
 			fmt.Printf("index1: %d; index2: %d; value: %s\n", index, index2, value2.ToString())
 		}
